circuit/builder: contains step0 keccak, needs testing, circuit/builder/word64: new file for Word64 logic and other keccak utils, Cargo.toml: added my name
